Enabled to pass parameters how to connect ACOS appliance
All actions can't connect an ACOS appliance when its SSL feature is disabled because
BaseAction that leads processing to handle acos.Client never pass parameters to
manupirate how to connect ACOS appliance.

This commit fixed it to be able to pass those parameters from the "specified_target"
parameter that is equipped every action.